Add files via upload
The FP multimode semiconductor laser has lightened up a new field of optical
communication technology in the past two decades. Numerical modeling of its
physical behaviours and transient response has been discussed previously in literature,
mostly by constructing the multimode rate equations. Rate equations are
very helpful in studying and predicting the average photon and carrier transient
response and relaxation oscillation. However, their deficiency in statistical photon
fluctuation limits the function of describing stochastic power shifted from main
mode to other side modes. Therefore, a noise driven model with conjunction of
optical fibre and photodiode is built to form an optical communication system in
the simulation scope.
The multimode nature of FP lasers causes several problems such as mode partitioning
noise (MPN), intersymbol interference (lSI), and frequency chirping, among
which mode partitioning noise is the most serious of the concern in this discussion.
The stereotype analytical measurement of MPN power penalty is based on several
assumptions on the received waveform shape and power distribution spectrum,
which limits its fields of application and accuracy. This work develops a numerical
solution to power penalty due to MPN, and it can be employed to any multimode
laser diode models regardless of the received signal shape and power distribution
In conclusion, the MPN power penalty is a significant profile of evaluating system
perform in fibre-optic communication links. It highly depends on shape of
power distribution spectrum, number of modes, length of fibre, and pattern of
signal waveform.
Conclusion and Future Works
This work establishes a complete numerical model of noise driven FP multimode
laser diode rate equations. With connecting to a single mode fibre and PIN photodiode,
it forms a basic optical communication system with laser as the transmitter,
fibre as the channel and photodiode as the receiver. The focus of problems
caused by multi wavelength transmission is to analyze the mode partitioning
effects, which can cause intersymbol interference, system power penalty, signal
distortion and etc. Additionally, a simple improvement is developed to critically
refine the system performance.
There are certainly some works can be done in the future. With absence of a optical
amplifier, the signal intensity is greatly reduced after going through the photodiode.
To complete the system, an optical amplifier is usually applied in reality.
However, it not only enlarges the signal but also the noise. In this case, developing
another method of improvement becomes a necessity, such as modification on
modulation scheme, application of an equalizer and a mode lock device.